+++ tg-catalog-logistics/src/main/java/com/tiendasgo/catalog/domain/repository/MarcaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MarcaRepository extends JpaRepository<Marca, Long> {
+public interface MarcaRepository extends JpaRepository<Marca, Integer> {
 
     Optional<Marca> findByNombre(String nombre);
