+++ tg-catalog-logistics/src/main/java/com/tiendasgo/catalog/domain/repository/SubCategoriaRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Long> {
+public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Integer> {
 
     List<SubCategoria> findByCategoriaId(Integer categoriaId);
